Add pause to wait for scroll stop in WPT pointerevent touch test
In order to run the web-platform-tests automatically, we will use
testdriver Action API in all the wpt to simulate inputs. Here we are
changing some pointerevent touch input tests which are doing scrolling.

We are usinig "pause" action in the action sequence to  wait for the
scroll to stop to avoid the memory leak.

Bug: 606367
Change-Id: Ic55144000c710789639e0996a3a3fab5ce3bdd42