Use matching sample rate for the context as for the reference file
The reference file has a certain sample rate.  Use the same for the
context to reduce the amount of resampling that might be required.

Regenerated the webm-decode-expected.wave file with the fixed sample rate.  This also
changed the SNR slightly, so update the SNR test accordingly.  Also manually compared
the new and old files.  These differ by at most 1 LSB (16-bit).

Bug: 940970
Change-Id: I979bbeb512c36d7e2eb4cba85550fa4177b8aabd