[css-properties-values-api] Compute <length> values to px.
According to the spec, the computed value of <length> registered properties
is the absolute value expressed in pixels. Currently, we are only behaving
correctly if the value is a calc() or if the CSSPrimitiveValue contains a
relative unit (e.g. em). This means that values like '1in' would not be
expressed in pixels.

As of this CL, we absolutize any <length> which isn't already a simple
(non-calc) px value.

R=futhark@chromium.org

Bug: 641877
Change-Id: Idec079671a094952e122847fc5ebd4b0cb609674